services/generator/requirements.txt
alembic==1.4.1
appdirs==1.4.4
atpublic==2.3
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
croniter==0.3.37
cycler==0.10.0
dask==2021.4.0
databricks-cli==0.14.3
decorator==4.4.2
dictdiffer==0.8.1
diskcache==5.2.1
distributed==2021.4.0
distro==1.5.0
docker==5.0.0
dpath==2.0.1
dulwich==0.20.21
dvc==2.0.18
entrypoints==0.3
Flask==1.1.2
flatten-dict==0.3.0
flufl.lock==3.2
fsspec==0.9.0
ftfy==6.0
funcy==1.15
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
grandalf==0.6
greenlet==1.0.0
gunicorn==20.1.0
HeapDict==1.0.1
idna==2.10
importlib-metadata==3.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonpath-ng==1.5.2
kiwisolver==1.3.1
locket==0.2.1
mailchecker==4.0.6
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-oneofschema==2.1.0
matplotlib==3.4.1
mlflow==1.15.0
msgpack==1.0.2
mypy-extensions==0.4.3
nanotime==0.5.2
natsort==7.1.1
networkx==2.5.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
partd==1.2.0
pathlib2==2.3.5
pathspec==0.8.1
pendulum==2.1.2
phonenumbers==8.12.21
Pillow==8.2.0
ply==3.11
prefect==0.14.16
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.1
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pycparser==2.20
pydot==1.4.2
pygit2==1.5.0
Pygments==2.8.1
pygtrie==2.3.2
pyparsing==2.4.7
python-benedict==0.23.2
python-box==5.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-fsutil==0.4.0
python-slugify==4.0.1
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
rich==10.1.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
shortuuid==1.0.1
shtab==1.3.5
six==1.15.0
smmap==4.0.0
sortedcontainers==2.3.0
SQLAlchemy==1.4.8
sqlparse==0.4.1
tabulate==0.8.9
tblib==1.7.0
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
voluptuous==0.12.1
wcwidth==0.2.5
websocket-client==0.58.0
Werkzeug==1.0.1
xmltodict==0.12.0
zc.lockfile==2.0
zict==2.0.0
zipp==3.4.1
